--- app/components/selection_table.py
@@ -1,5 +1,0 @@
-from dash import html
-
-selection_table = html.Div(
-        id='well-selection-table'
-    )
