--- app/components/tabs_content.py
@@ -1,30 +1,0 @@
-########################################################################
-####                                                                ####
-####                             Imports                            ####
-####                                                                ####
-########################################################################
-
-import dash_bootstrap_components as dbc
-from dash import dcc
-
-from app.pages.configure import configure_analysis
-from app.pages.metadata import meta_data
-########################################################################
-####                                                                ####
-####                             Layout                             ####
-####                                                                ####
-########################################################################
-# Tabs Content
-tabs_content = dbc.Container(
-    [
-        dcc.Tabs(id="tabs", value='video_analysis-tab', children=[
-            dcc.Tab(label="Configure Analysis", value="video_analysis-tab", children=[
-                configure_analysis
-            ]),
-            dcc.Tab(label="Metadata", value='metadata-tab', children=[
-                meta_data
-            ])
-        ])
-    ],
-    className="mt-5"
-)
