+++ app/pages/metadata.py
@@ -7,13 +7,14 @@ import dash_bootstrap_components as dbc
 import dash
 from dash.dependencies import Input, Output, State
 from dash import dcc, html, callback, dash_table
+import pandas as pd
+from pathlib import Path
 
 # components
-from app.components.metadata_table_checklist import metadata_checklist
-from app.components.create_metadata_tabs_from_checklist import meta_data_from_input
+from app.components.metadata_components import metadata_checklist
 
 # utilities
-from app.utils.create_df_from_user_input import create_df_from_inputs, create_empty_df_from_inputs
+from app.utils.callback_functions import create_empty_df_from_inputs
 
 dash.register_page(__name__)
 
@@ -104,108 +105,11 @@ layout = dbc.Container(
     style={"paddingTop": "80px"}
 )
 
-
-@callback(
-    [Output("table-container-batch", "children"),
-     Output("table-container-species", "children"),
-     Output("table-container-strains", "children"),
-     Output("table-container-stages", "children"),
-     Output("table-container-treatments", "children"),
-     Output("table-container-conc", "children"),
-     Output("table-container-other", "children"),
-     Output("well-selection-table", 'children')
-     ],
-    State('store', 'data'),
-    [Input("total-num-rows", "value"),
-     Input("total-well-cols", "value")]
-)
-def update_table(store, rows, cols):
-    default_cols = 12
-    default_rows = 8
-    try:
-        num_rows = store['rows']
-    except:
-        num_rows = default_rows
-    try:
-        num_cols = store['cols']
-    except:
-        num_cols = default_cols
-
-    df = create_df_from_inputs(rows, cols)
-    df_empty = create_empty_df_from_inputs(rows, cols)
-    table_batch = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-batch'
-    )
-    table_species = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-species'
-    )
-    table_stages = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-stages'
-    )
-    table_strains = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-strains'
-    )
-    table_treatment = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-species'
-    )
-    table_conc = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-conc'
-    )
-    table_other = dash_table.DataTable(
-        data=df_empty.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-other'
-    )
-    well_selection = dash_table.DataTable(
-        data=df.reset_index().to_dict('records'),
-        columns=[{'name': 'Row', 'id': 'index', 'editable': False}] +
-        [{'name': col, 'id': col} for col in df.columns],
-        editable=True,
-        style_table={'overflowX': 'auto'},
-        style_cell={'textAlign': 'center'},
-        id='dynamic-table-container-well-selection-table'
-    )
-    return table_batch, table_species, table_strains, table_stages, table_treatment, table_conc, table_other, well_selection
+########################################################################
+####                                                                ####
+####                             Callbacks                          ####
+####                                                                ####
+########################################################################
 
 
 @callback(
@@ -252,3 +156,62 @@ def create_tabs_from_checklist(store, n_clicks, checklist_values):
     else:
         # If no checklist values are available, return an empty list and set 'batch-data-tab' as the selected tab
         return [], 'batch-data-tab'
+
+
+@callback(
+    Output("checklist-input", "options"),
+    [Input("add-metadata-table-button", "n_clicks")],
+    [State("uneditable-input-box", 'value'),
+     State("checklist-input", "options")]
+)
+def update_metadata_checklist(n_clicks, new_table_name, existing_options):
+    if n_clicks and new_table_name:
+        # Append the new table name to the existing options
+        new_option = {"label": new_table_name, "value": new_table_name}
+        updated_options = existing_options + [new_option]
+        return updated_options
+    else:
+        return existing_options
+
+
+@callback(
+    Output("save-meta-data-to-csv", 'color'),
+    Input("save-meta-data-to-csv", "n_clicks"),
+    State('metadata-tabs', 'children'),
+    State('mounted-volume', 'value')
+)
+def save_the_metadata_tables_to_csv(n_clicks, metadata_tabs, volume):
+    if n_clicks:
+
+        # Iterate over the metadata tabs
+        for tab in metadata_tabs:
+            tab_data = tab['props']['children'][0]['props']['children']['props']['data']
+            tab_id = tab['props']['label']
+            tab_id = tab_id.lower()
+
+            df = pd.DataFrame(tab_data)
+            current_columns_order = df.columns.tolist()
+            # Define a mapping of column names to integer values (except for 'index')
+            column_order_mapping = {'index': -1}
+            for i, col in enumerate(current_columns_order):
+                if col != 'index':
+                    column_order_mapping[col] = int(col)
+
+            # Sort the columns based on their values in the mapping
+            sorted_columns = sorted(
+                current_columns_order, key=lambda col: column_order_mapping[col])
+            sorted_columns = list(sorted_columns)
+            sorted_columns = sorted_columns[1:]
+
+            # Reorder the DataFrame columns
+            df = df[sorted_columns]
+
+            # Save the DataFrame to a CSV file
+            metadata_dir = Path(volume).joinpath('metadata')
+            if not metadata_dir.exists():
+                metadata_dir.mkdir(parents=True, exist_ok=True)
+            file_path = metadata_dir.joinpath(f"{tab_id}.csv")
+            df.to_csv(file_path, index=False, header=False)
+
+    # Enable the button if not clicked
+    return "success"
