+++ app/pages/preview.py
@@ -5,7 +5,7 @@
 ########################################################################
 import dash_bootstrap_components as dbc
 import dash
-from dash import dcc, html
+from dash import dcc, html, callback
 from dash.dependencies import Input, Output, State
 import docker
 import yaml
@@ -79,33 +79,6 @@ layout = dbc.ModalBody(
     ],
 )
 
-########################################################################
-####                                                                ####
-####                             Modal                              ####
-####                                                                ####
-########################################################################
-# Preview Page Modal
-preview_page = dbc.Modal(
-    [
-        # Modal Header
-        dbc.ModalHeader(
-            "Preview Page"
-        ),
-        # Modal page contents from defined above
-        layout,
-
-        # Modal Footer
-        dbc.ModalFooter([
-            # Buttons for the Info Page Modal
-            dbc.Button("Analyze all videos", id="run-button",
-                       className="ml-auto", color="success"),
-            dbc.Button("Close", id="close-preview-modal", className="ml-auto"),
-        ]),
-        html.Div(id="preview-page-status")
-    ],
-    id="preview-page-modal",
-    size="xl"
-)
 
 ########################################################################
 ####                                                                ####
@@ -114,141 +87,159 @@ preview_page = dbc.Modal(
 ########################################################################
 
 
-def load_first_img(app):
-    @app.callback(
-        Output('input-path-output', 'children'),
-        Output('input-preview', 'figure'),
-        Input('submit-val', 'n_clicks'),
-        State('mounted-volume', 'value'),
-        State('plate-name', 'value'),
-        State('well-selection-list', 'children'),
-        prevent_initial_call=True
-    )
-    def update_preview_image(n_clicks, volume, platename, wells):
+@callback(
+    Output('input-path-output', 'children'),
+    Output('input-preview', 'figure'),
+    Input('submit-val', 'n_clicks'),
+    State('mounted-volume', 'value'),
+    State('plate-name', 'value'),
+    State('well-selection-list', 'children'),
+    prevent_initial_call=True
+)
+def update_preview_image(n_clicks, volume, platename, wells):
 
-        first_well = wells[0].replace(', ', '')
+    first_well = wells[0].replace(', ', '')
 
-        plate_base = platename.split("_", 1)[0]
+    plate_base = platename.split("_", 1)[0]
 
-        if n_clicks >= 1:
-            # assumes IX-like file structure
-            img_path = Path(
-                volume, 'input', f'{platename}/TimePoint_1/{plate_base}_{first_well}.TIF')
-            img = np.array(Image.open(img_path))
-            fig = px.imshow(img, color_continuous_scale="gray")
-            fig.update_layout(coloraxis_showscale=False)
-            fig.update_xaxes(showticklabels=False)
-            fig.update_yaxes(showticklabels=False)
-            return f'```{img_path}```', fig
-        n_clicks = 0
-
-
-def populate_options(app):
-    @app.callback(
-        Output('preview-dropdown', 'options'),
-        Input('open-preview-modal', 'n_clicks'),
-        State('motility-run', 'value'),
-        State('segment-run', 'value')
-    )
-    def get_options(nclicks, motility, segment):
-
-        selection_dict = {'motility': 'motility', 'segment': 'binary'}
-        option_dict = {}
-
-        if nclicks is not None:
-            for selection in selection_dict.keys():
-                if eval(selection) == 'True':
-                    option_dict[selection] = selection_dict[selection]
-
-        dict_option = {v: k for k, v in option_dict.items()}
-        return dict_option
+    if n_clicks >= 1:
+        # assumes IX-like file structure
+        img_path = Path(
+            volume, 'input', f'{platename}/TimePoint_1/{plate_base}_{first_well}.TIF')
+        img = np.array(Image.open(img_path))
+        fig = px.imshow(img, color_continuous_scale="gray")
+        fig.update_layout(coloraxis_showscale=False)
+        fig.update_xaxes(showticklabels=False)
+        fig.update_yaxes(showticklabels=False)
+        return f'```{img_path}```', fig
+    n_clicks = 0
 
 
-def preview_analysis(app):
-    @app.callback(
-        Output('analysis-preview-message', 'children'),
-        Output('analysis-preview', 'figure'),
-        Input('preview-button', 'n_clicks'),
-        State('imaging-mode', 'value'),
-        State('file-structure', 'value'),
-        State('multi-well-rows', 'value'),
-        State('multi-well-cols', 'value'),
-        State('multi-well-detection', 'value'),
-        State('species', 'value'),
-        State('stages', 'value'),
-        State('motility-run', 'value'),
-        State('conversion-run', 'value'),
-        State('conversion-scale-video', 'value'),
-        State('conversion-rescale-multiplier', 'value'),
-        State('segment-run', 'value'),
-        State('segmentation-wavelength', 'value'),
-        State('cell-profiler-run', 'value'),
-        State('cell-profiler-pipeline', 'value'),
-        State('diagnostics-dx', 'value'),
-        State('plate-name', 'value'),
-        State('mounted-volume', 'value'),
-        State('well-selection-list', 'children'),
-        State('preview-dropdown', 'value'),
-        prevent_initial_call=True
-    )
-    def run_analysis(nclicks, imagingmode, filestructure, multiwellrows, multiwellcols, multiwelldetection, species, stages, motilityrun, conversionrun, conversionscalevideo, conversionrescalemultiplier, segmentrun, wavelength, cellprofilerrun, cellprofilerpipeline, diagnosticdx, platename, volume, wells, selection):
-        if nclicks:
+@callback(
+    Output('preview-dropdown', 'options'),
+    Input('open-preview-modal', 'n_clicks'),
+    State('motility-run', 'value'),
+    State('segment-run', 'value')
+)
+def get_options(nclicks, motility, segment):
 
-            if wells == 'All':
-                first_well = 'A01'
-            else:
-                first_well = wells[0]
+    selection_dict = {'motility': 'motility', 'segment': 'binary'}
+    option_dict = {}
 
-            config = prep_yaml(
-                imagingmode,
-                filestructure,
-                multiwellrows,
-                multiwellcols,
-                multiwelldetection,
-                species,
-                stages,
-                motilityrun,
-                conversionrun,
-                conversionscalevideo,
-                conversionrescalemultiplier,
-                segmentrun,
-                wavelength,
-                cellprofilerrun,
-                cellprofilerpipeline,
-                diagnosticdx,
-                wells
-            )
+    if nclicks is not None:
+        for selection in selection_dict.keys():
+            if eval(selection) == 'True':
+                option_dict[selection] = selection_dict[selection]
 
-            output_file = Path(volume, platename + '.yml')
+    dict_option = {v: k for k, v in option_dict.items()}
+    return dict_option
 
-            # Dump preview data to YAML file
-            with open(output_file, 'w') as yaml_file:
-                yaml.dump(config, yaml_file,
-                          default_flow_style=False)
 
-            client = docker.from_env()
-            print(client)
+@callback(
+    Output('analysis-preview-message', 'children'),
+    Output('analysis-preview', 'figure'),
+    Input('preview-button', 'n_clicks'),
+    State('imaging-mode', 'value'),
+    State('file-structure', 'value'),
+    State('multi-well-rows', 'value'),
+    State('multi-well-cols', 'value'),
+    State('multi-well-detection', 'value'),
+    State('species', 'value'),
+    State('stages', 'value'),
+    State('motility-run', 'value'),
+    State('conversion-run', 'value'),
+    State('conversion-scale-video', 'value'),
+    State('conversion-rescale-multiplier', 'value'),
+    State('segment-run', 'value'),
+    State('segmentation-wavelength', 'value'),
+    State('cell-profiler-run', 'value'),
+    State('cell-profiler-pipeline', 'value'),
+    State('diagnostics-dx', 'value'),
+    State('plate-name', 'value'),
+    State('mounted-volume', 'value'),
+    State('well-selection-list', 'children'),
+    State('preview-dropdown', 'value'),
+    prevent_initial_call=True
+)
+def run_analysis(nclicks,
+                 imagingmode,
+                 filestructure,
+                 multiwellrows,
+                 multiwellcols,
+                 multiwelldetection,
+                 species,
+                 stages,
+                 motilityrun,
+                 conversionrun,
+                 conversionscalevideo,
+                 conversionrescalemultiplier,
+                 segmentrun,
+                 wavelength,
+                 cellprofilerrun,
+                 cellprofilerpipeline,
+                 diagnosticdx,
+                 platename,
+                 volume,
+                 wells,
+                 selection
+                 ):
+    if nclicks:
 
-            command = f"python wrmXpress/wrapper.py {platename}.yml {platename}"
-            command_message = f"```python wrmXpress/wrapper.py {platename}.yml {platename}```"
+        if wells == 'All':
+            first_well = 'A01'
+        else:
+            first_well = wells[0]
 
-            container = client.containers.run('zamanianlab/wrmxpress', command=f"{command}", detach=True,
-                                              volumes={f'{volume}/input/': {'bind': '/input/', 'mode': 'rw'},
-                                                       f'{volume}/output/': {'bind': '/output/', 'mode': 'rw'},
-                                                       f'{volume}/work/': {'bind': '/work/', 'mode': 'rw'},
-                                                       f'{volume}/{platename}.yml': {'bind': f'/{platename}.yml', 'mode': 'rw'}
-                                                       })
+        config = prep_yaml(
+            imagingmode,
+            filestructure,
+            multiwellrows,
+            multiwellcols,
+            multiwelldetection,
+            species,
+            stages,
+            motilityrun,
+            conversionrun,
+            conversionscalevideo,
+            conversionrescalemultiplier,
+            segmentrun,
+            wavelength,
+            cellprofilerrun,
+            cellprofilerpipeline,
+            diagnosticdx,
+            wells
+        )
 
-            # assumes IX-like file structure
-            img_path = Path(
-                volume, 'work', f'{platename}/{first_well}/img/{platename}_{first_well}_{selection}.png')
+        output_file = Path(volume, platename + '.yml')
 
-            while not os.path.exists(img_path):
-                time.sleep(1)
+        # Dump preview data to YAML file
+        with open(output_file, 'w') as yaml_file:
+            yaml.dump(config, yaml_file,
+                      default_flow_style=False)
 
-            img = np.array(Image.open(img_path))
-            fig = px.imshow(img, color_continuous_scale="gray")
-            fig.update_layout(coloraxis_showscale=False)
-            fig.update_xaxes(showticklabels=False)
-            fig.update_yaxes(showticklabels=False)
-            return command_message, fig
+        client = docker.from_env()
+        print(client)
+
+        command = f"python wrmXpress/wrapper.py {platename}.yml {platename}"
+        command_message = f"```python wrmXpress/wrapper.py {platename}.yml {platename}```"
+
+        container = client.containers.run('zamanianlab/wrmxpress', command=f"{command}", detach=True,
+                                          volumes={f'{volume}/input/': {'bind': '/input/', 'mode': 'rw'},
+                                                   f'{volume}/output/': {'bind': '/output/', 'mode': 'rw'},
+                                                   f'{volume}/work/': {'bind': '/work/', 'mode': 'rw'},
+                                                   f'{volume}/{platename}.yml': {'bind': f'/{platename}.yml', 'mode': 'rw'}
+                                                   })
+
+        # assumes IX-like file structure
+        img_path = Path(
+            volume, 'work', f'{platename}/{first_well}/img/{platename}_{first_well}_{selection}.png')
+
+        while not os.path.exists(img_path):
+            time.sleep(1)
+
+        img = np.array(Image.open(img_path))
+        fig = px.imshow(img, color_continuous_scale="gray")
+        fig.update_layout(coloraxis_showscale=False)
+        fig.update_xaxes(showticklabels=False)
+        fig.update_yaxes(showticklabels=False)
+        return command_message, fig
