--- app/utils/create_df_from_user_input.py
@@ -1,28 +1,0 @@
-########################################################################
-####                                                                ####
-####                             Imports                            ####
-####                                                                ####
-########################################################################
-import pandas as pd
-
-
-########################################################################
-####                                                                ####
-####                             LAYOUT                             ####
-####                                                                ####
-########################################################################
-def create_df_from_inputs(_rows, _cols):
-    rows_total = list("ABCDEFGHIJKLMNOP")
-    rows = rows_total[:int(_rows)]
-    columns = [str(num).zfill(2) for num in range(1, int(_cols) + 1)]
-    data = [[row + col for col in columns] for row in rows]
-    df = pd.DataFrame(data, columns=columns, index=rows)
-    return df
-
-def create_empty_df_from_inputs(_rows, _cols):
-    rows_total = list("ABCDEFGHIJKLMNOP")
-    rows = rows_total[:int(_rows)]
-    columns = [str(num).zfill(2) for num in range(1, int(_cols) + 1)]
-    empty_data = [[None for _ in columns] for _ in rows]
-    df = pd.DataFrame(empty_data, columns=columns, index=rows)
-    return df
